ai-agents/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
fhirpy==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
aiohttp==3.9.5